Fold /account/debug into the Chancellor tools page
The impersonate form and the raw account debug dump move to
/account/settings/chancellor, alongside the other Chancellor-gated
admin tools; the old /account/debug route permanently redirects there.
The debug dump was never useful to non-Chancellors, and impersonation
was already Chancellor-gated, so everything admin now lives on one page.
