+++ CLAUDE.md
@@ -110,7 +110,7 @@ Functions take `(accessToken, id, ...)`; return raw ESI JSON (paged wrappers ret
 
 ## App routes → files
 
-Routes follow the App Router convention: `/<path>` → `src/app/<path>/page.tsx` (or `route.ts` for handlers). Pages: `/`, `/account/{login,register,settings,invite,reset,email,debug}`, `/account/settings/chancellor`, `/account/gice/complete`, `/asset`, `/asset/[locationId]`, `/asset/search`, `/ship/[itemId]`, `/character`, `/jobs`, `/market`, `/industry`, `/indexes`, `/mercenary-dens`, `/corpses/[characterID]`, `/structure`, `/structure/revenue`, `/structure/[structureId]`, `/settings/grants`, `/settings/tax`, `/blueprint`, `/blueprint/[typeID]`, `/fitting`, `/fitting/[characterId]/[fittingId]`, `/oauth/consent`. Handlers: `/account/confirm`, `/account/gice` (+`/callback`), `/character/callback`, `/esf/[file]`, `/sheets/[file]`, `/sheets/market/[market]`, `/api/mcp`, `/.well-known/oauth-protected-resource`, `/api/{character,corp}/*` (CSV), `/api/fittings/*` (the fittings archive; a restore is `PUT` to a client-minted uuid, since EVE assigns the fitting id itself), `/api/type/search`, `/xrpc/[method]`.
+Routes follow the App Router convention: `/<path>` → `src/app/<path>/page.tsx` (or `route.ts` for handlers). Pages: `/`, `/account/{login,register,settings,invite,reset,email}`, `/account/settings/chancellor`, `/account/gice/complete`, `/asset`, `/asset/[locationId]`, `/asset/search`, `/ship/[itemId]`, `/character`, `/jobs`, `/market`, `/industry`, `/indexes`, `/mercenary-dens`, `/corpses/[characterID]`, `/structure`, `/structure/revenue`, `/structure/[structureId]`, `/settings/grants`, `/settings/tax`, `/blueprint`, `/blueprint/[typeID]`, `/fitting`, `/fitting/[characterId]/[fittingId]`, `/oauth/consent`. Handlers: `/account/confirm`, `/account/gice` (+`/callback`), `/character/callback`, `/esf/[file]`, `/sheets/[file]`, `/sheets/market/[market]`, `/api/mcp`, `/.well-known/oauth-protected-resource`, `/api/{character,corp}/*` (CSV), `/api/fittings/*` (the fittings archive; a restore is `PUT` to a client-minted uuid, since EVE assigns the fitting id itself), `/api/type/search`, `/xrpc/[method]`.
 
 Notable pages:
 
@@ -213,7 +213,7 @@ Key Postgres functions (RPC or SQL):
 
 - **SCD Type 2:** `is_current=true` rows form the current snapshot; `valid_until` is bumped for unchanged rows; a change inserts a new row and closes the old one. Applied to character/corp assets, blueprints, ship (single open row per character), skills (keyed `(registration_id, skill_id)`; never unlearned, so superseded, never closed for vanishing), orders/industry jobs (keyed `order_id`/`job_id`), fittings (keyed `(registration_id, fitting_id)`; an in-client edit opens a new version, a delete closes the row). Vanishing-row semantics differ: a vanished order has filled/expired/cancelled → closed (view holds only open orders); a vanished industry job aged past ESI's `include_completed` window → a **terminal** row (`delivered`/`cancelled`/`reverted`) stays `is_current` (view keeps every job we saw finish), a **non-terminal** one is closed. That window runs ~90 days from _install_, not from delivery, so a job outlasting it (multi-month ME research) vanishes before its `delivered` state is ever observable; leaving those pinned a stale `active` nothing could supersede. The split is the pure `partitionJobs` seam (`src/jobs/industryJobReconcile.js`, shared by both job modules, tested), which assumes a **complete** listing — one response for character (no `x-pages`), `fetchAllPages` for corp.
 - **Asset location-walk functions, two shapes:** the `summary`/`contents` functions seed from every asset (whole-hangar aggregate); the `search` functions seed from only the rows matching a filter, so search stays cheap — reuse the seeded-recursion shape for future "look up a few items, walk their tree" functions.
-- **Chancellor (admin) accounts:** redeemed an `is_chancellor` invite code. `isChancellor(userId)` checks via service role (RLS only shows a user codes they created). Gates `/account/settings/chancellor`, the impersonate form on `/account/debug`, and on-demand `industry-systems`.
+- **Chancellor (admin) accounts:** redeemed an `is_chancellor` invite code. `isChancellor(userId)` checks via service role (RLS only shows a user codes they created). Gates `/account/settings/chancellor` (which carries the impersonate form and the account debug dump; old `/account/debug` redirects here) and on-demand `industry-systems`.
 - **Sheets IMPORTDATA:** the `/api/character/*` and `/api/corp/*` routes authenticate via `user_settings.api_token`, call a Postgres function, return CSV.
 - **The one ESI write path** (`/api/fittings`, docs/fitting-fuse.md): everywhere else is UI → DB → extract job → ESI, read-only. Archiving a fitting deletes it in game, so `esi-fittings.write_fittings.v1` is the first write scope — carried by the new `optIn` flag on `EsiScope`, which keeps a scope out of `defaultScopes` (opt-in, unlike every other optional scope, which is on by default). The writes live only in `src/app/api/fittings/lib.ts` and are **synchronous, not queued**: the caller is a filesystem, so `rm` must have happened before it returns. Every write inserts its whole fit body into `fitting_write_log` _before_ calling ESI (a crash leaves a `pending` row holding a restorable copy; a fit deleted with nothing stored cannot happen), fetches live from ESI first (the mirror can be 6h stale), and converges the SCD rows itself so the next listing agrees. Identity is `contentHash` (src/fittingArchive.ts), never `fitting_id` — the game reassigns that on restore. A restore is addressed by a caller-minted uuid (`fitting_write_log.request_id`) so it can be a `PUT`; the key is **used once** (a reuse is `409`, enforced by a unique index rather than the read before it — `test/sql/fitting_write_log.sql`), while the content hash catches the same fit arriving under a fresh key.
 - **Vercel queue:** only the innomin.at appraisal throttle still uses it (topic `innominate`, consumer `/api/queue/innominate`).
